alter tsv color order
Alter the order for tsv colors based on subjective testing on a dark background.

Repeats a complementary cycle of YBRGM, light and then dark, with cyan as light blue and white as light yellow. Starts with white for the sake of non-tabular lines (e.g. textual comments); this uniquely determines the sequence up to the choice of cyan or magenta as column 2. I chose cyan.